Fixed a few errors with inventories

+++ src/main/java/com/uberverse/arkcraft/common/entity/data/ARKPlayer.java
@@ -209,7 +209,7 @@ public class ARKPlayer implements IExtendedEntityProperties
 	
 	public double getCarryWeightRatio() 
 	{
-		return (double) carryWeight / weight;
+		return (double) carryWeight /  (double) weight;
 	}
 
 	/**
